mcpp (2.6.1-1) unstable; urgency=low

  * configure.ac: Fix for automake 1.9.
  * debian/rules: Install the manual.
  * debian/control (Standards-Version): Follows new standard 3.7.2.2.
  * debian/watch: The upstram has been moved.
  * New upstream release

+++ src/set_mcpp.sh
@@ -13,6 +13,8 @@ inc_dir=$9
 cpp_name=`echo ${cpp_call} | sed 's,.*/,,'`
 cpp_path=`echo ${cpp_call} | sed "s,/${cpp_name},,"`
 gcc_path=`echo $1 | sed "s,/${CC}\$,,"`
+sys_mingw=`echo $1 | grep 'mingw'`
+sys_cygwin=`echo ${inc_dir} | grep 'cygwin'`
 
 # remove ".exe" or such
 EXEEXT=`echo $7 | sed 's/^x//'`
@@ -22,52 +24,103 @@ else
     cpp_base=${cpp_name}
 fi
 
-echo "  cp -f mcpp_g*${gcc_maj_ver}${gcc_min_ver}_predef_*.h ${inc_dir}"
-cp -f mcpp_g*${gcc_maj_ver}${gcc_min_ver}_predef_*.h ${inc_dir}
+if test ${sys_mingw} && test ! -f cc1${EXEEXT}; then
+    ## cc1.exe has not yet compiled
+    echo "  do 'make COMPILER=GNUC mcpp cc1'; then do 'make COMPILER=GNUC install'"
+    exit 1
+fi
+
+cwd=`pwd`
+
+echo "  cd ${inc_dir}"
+cd ${inc_dir}
+if test ! -f mcpp_gcc${gcc_maj_ver}${gcc_min_ver}_predef_std.h; then
+    echo "  generating mcpp_g*.h header files"
+    echo '' | ${CC} -E -xc -dM - | sort | grep ' *#define *_'       \
+            > mcpp_gcc${gcc_maj_ver}${gcc_min_ver}_predef_std.h
+    echo '' | ${CC} -E -xc -dM - | sort | grep -E ' *#define *[A-Za-z]+'    \
+            > mcpp_gcc${gcc_maj_ver}${gcc_min_ver}_predef_old.h
+    echo '' | ${CXX} -E -xc++ -dM - | sort | grep ' *#define *_'    \
+            > mcpp_gxx${gcc_maj_ver}${gcc_min_ver}_predef_std.h
+    echo '' | ${CXX} -E -xc++ -dM - | sort | grep -E ' *#define *[A-Za-z]+' \
+            > mcpp_gxx${gcc_maj_ver}${gcc_min_ver}_predef_old.h
+fi
+if test ${sys_cygwin}; then
+    mkdir -p mingw
+    cd mingw
+    if test ! -f mcpp_gcc${gcc_maj_ver}${gcc_min_ver}_predef_std.h; then
+        echo "  generating mcpp_g*.h header files for cygwin/mingw"
+        echo '' | ${CC} -E -xc -dM -mno-cygwin - | sort |   \
+                grep ' *#define *_'       \
+                > mcpp_gcc${gcc_maj_ver}${gcc_min_ver}_predef_std.h
+        echo '' | ${CC} -E -xc -dM -mno-cygwin - | sort |   \
+                grep -E ' *#define *[A-Za-z]+'    \
+                > mcpp_gcc${gcc_maj_ver}${gcc_min_ver}_predef_old.h
+        echo '' | ${CXX} -E -xc++ -dM -mno-cygwin - | sort |    \
+                grep ' *#define *_'    \
+                > mcpp_gxx${gcc_maj_ver}${gcc_min_ver}_predef_std.h
+        echo '' | ${CXX} -E -xc++ -dM -mno-cygwin - | sort |    \
+                grep -E ' *#define *[A-Za-z]+' \
+                > mcpp_gxx${gcc_maj_ver}${gcc_min_ver}_predef_old.h
+    fi
+fi
 
 # write shell-script to call of 'cpp0', 'cc1 -E' or so is replaced to call of
-# mcpp_std
+# mcpp
 echo "  cd ${cpp_path}"
 cd ${cpp_path}
-echo '#!/bin/sh'        >  mcpp.sh
 
-# for GNU C V.2, V.3.3 and later
-if test x${gcc_maj_ver} = x2 || test x${cpp_base} = xcc1; then
-    echo 'for i in $@'  >> mcpp.sh
-    echo '    do'       >> mcpp.sh
-    echo '    case $i in'           >> mcpp.sh
-    echo '        -traditional*| -fpreprocessed)'           >> mcpp.sh
-    echo "            ${cpp_path}/${cpp_base}_gnuc"' "$@"'  >> mcpp.sh
-    echo '            exit ;;'      >> mcpp.sh
-    echo '    esac'     >> mcpp.sh
-    echo 'done'         >> mcpp.sh
+# other than MinGW
+if test ! ${sys_mingw}; then
+
+    echo '#!/bin/sh'        >  mcpp.sh
+
+    # for GCC V.3.3 and later
+    if test x${cpp_base} = xcc1; then
+        cat >> mcpp.sh <<_EOF
+for i in \$@
+    do
+    case \$i in
+        -fpreprocessed)
+            ${cpp_path}/${cpp_base}_gnuc "\$@"
+            exit ;;
+    esac
+done
+_EOF
+        cat > mcpp_plus.sh <<_EOF
+#!/bin/sh
+for i in \$@
+    do
+    case \$i in
+        -fpreprocessed)
+            ${cpp_path}/cc1plus_gnuc "\$@"
+            exit ;;
+    esac
+done
+_EOF
+    fi
+    
+    # for GCC V.2, V.3 and V.4
+    echo ${cpp_path}/mcpp '"$@"'   >>  mcpp.sh
+    chmod a+x mcpp.sh
+    if test x${cpp_base} = xcc1; then
+        echo ${cpp_path}/mcpp -+ '"$@"'  >> mcpp_plus.sh
+        chmod a+x mcpp_plus.sh
+    fi
 fi
 
-# for GNU C V.3.3 and later
-if test x${cpp_base} = xcc1; then
-    echo '#!/bin/sh'    >  mcpp_plus.sh
-    echo 'for i in $@'  >> mcpp_plus.sh
-    echo '    do'       >> mcpp_plus.sh
-    echo '    case $i in'           >> mcpp_plus.sh
-    echo '        -traditional*| -fpreprocessed)'       >> mcpp_plus.sh
-    echo "            ${cpp_path}/cc1plus_gnuc"' "$@"'  >> mcpp_plus.sh
-    echo '            exit ;;'      >> mcpp_plus.sh
-    echo '    esac'     >> mcpp_plus.sh
-    echo 'done'         >> mcpp_plus.sh
-fi
-
-# for GNU C V.2 and V.3
-echo ${cpp_path}/mcpp_std -23j '"$@"'   >>  mcpp.sh
-chmod a+x mcpp.sh
-if test x${cpp_base} = xcc1; then
-    echo ${cpp_path}/mcpp_std -+23j '"$@"'  >> mcpp_plus.sh
-    chmod a+x mcpp_plus.sh
-fi
-
-# backup GNU C / cpp or cc1, cc1plus
+# backup GCC / cpp or cc1, cc1plus
 if test `echo '' | ${cpp_call} -v - 2>&1 | grep 'MCPP' > /dev/null; echo $?`;
         then
-    sym_link=`ls -l ${cpp_name} | sed 's/^l.*/l/; s/^[^l].*//'`
+    if test ${sys_mingw}; then
+        if test -f cc1_gnuc${EXEEXT}; then
+            sym_link=l          ## cc1.exe already moved to cc1_gnuc.exe
+        else
+            sym_link=
+        fi
+    else
+        sym_link=`ls -l ${cpp_name} | sed 's/^l.*/l/; s/^[^l].*//'`
+    fi
     if test x${sym_link} != xl; then
         echo "  mv ${cpp_name} ${cpp_base}_gnuc${EXEEXT}"
         mv -f ${cpp_name} ${cpp_base}_gnuc${EXEEXT}
@@ -85,28 +138,42 @@ if test -f ${cpp_name}; then
 fi
 
 # make symbolic link of mcpp.sh to 'cpp0' or 'cc1', 'cc1plus'
-echo "  ${LN_S} mcpp.sh ${cpp_name}"
-${LN_S} mcpp.sh ${cpp_name}
+if test ${sys_mingw}; then
+    echo "  cp ${cwd}/cc1${EXEEXT}"
+    cp ${cwd}/cc1${EXEEXT} .
+    strip cc1${EXEEXT}
+else
+    echo "  ${LN_S} mcpp.sh ${cpp_name}"
+    ${LN_S} mcpp.sh ${cpp_name}
+fi
 if test x${cpp_base} = xcc1; then
-    echo "  ${LN_S} mcpp_plus.sh cc1plus${EXEEXT}"
-    ${LN_S} mcpp_plus.sh cc1plus${EXEEXT}
+    if test ${sys_mingw}; then
+        echo "  cp cc1${EXEEXT} cc1plus${EXEEXT}"
+        cp cc1${EXEEXT} cc1plus${EXEEXT}
+    else
+        echo "  ${LN_S} mcpp_plus.sh cc1plus${EXEEXT}"
+        ${LN_S} mcpp_plus.sh cc1plus${EXEEXT}
+    fi
 fi
 
 if test x${gcc_maj_ver} = x2; then
     exit 0
 fi
 
-# for GNU C V.3 make ${CC}.sh and ${CXX}.sh to add -no-integrated-cpp option
+# for GCC V.3 or V.4 make ${CC}.sh and ${CXX}.sh to add -no-integrated-cpp
+# option
 echo "  cd ${gcc_path}"
 cd ${gcc_path}
 
-sym_link=`ls -l ${CC}${EXEEXT} | sed 's/^l.*/l/; s/^[^l].*//'`
-if test x${sym_link} = xl; then     # symbolic link
-    clink=`ls -L -i ${CC}${EXEEXT}`                         # dereference
-    inode=`echo ${clink} | sed 's/^ *//' | sed 's/ .*//'`
-    c_entity=`ls -i | grep ${inode} | sed '1p; 1,$d' | sed 's/^[ 0-9]*//'   \
-            | sed 's/\*$//'`
-    if test x${c_entity} = x${CC}.sh; then
+if test ! ${sys_mingw}; then
+    ref=${CC}${EXEEXT}
+    while ref=`readlink ${ref}`
+    do
+        c_entity=${ref};
+    done
+fi
+if test x${c_entity} != x; then     # symbolic linked file dereferenced
+    if test ${c_entity} = ${CC}.sh; then    # gcc.sh already installed
         exit 0
     fi
     if test x${EXEEXT} != x; then
@@ -114,21 +181,23 @@ if test x${sym_link} = xl; then     # symbolic link
     else
         c_entity_base=${c_entity}
     fi
-    echo "  mv ${c_entity} ${c_entity_base}_proper${EXEEXT}"
-    mv -f ${c_entity} ${c_entity_base}_proper${EXEEXT}
 else                                # not symbolic link
-    echo "  mv ${CC}${EXEEXT} ${CC}_proper${EXEEXT}"
-    mv -f ${CC}${EXEEXT} ${CC}_proper${EXEEXT}
-    c_entity=${CC}${EXEEXT}
+    if test ! ${sys_mingw} || test ! -f ${CC}_proper${EXEEXT}; then
+        echo "  mv ${CC}${EXEEXT} ${CC}_proper${EXEEXT}"
+        mv -f ${CC}${EXEEXT} ${CC}_proper${EXEEXT}
+    fi
+    c_entity_base=${gcc_path}/${CC}_proper
 fi
 
-sym_link=`ls -l ${CXX}${EXEEXT} | sed 's/^l.*/l/; s/^[^l].*//'`
-if test x${sym_link} = xl; then     # symbolic link
-    clink=`ls -L -i ${CXX}${EXEEXT}`
-    inode=`echo ${clink} | sed 's/^ *//' | sed 's/ .*//'`
-    cxx_entity=`ls -i | grep ${inode} | sed '1p; 1,$d' | sed 's/^[ 0-9]*//' \
-            | sed 's/\*$//'`
-    if test x${cxx_entity} = x${CXX}.sh; then
+if test ! ${sys_mingw}; then
+    ref=${CXX}${EXEEXT}
+    while ref=`readlink ${ref}`
+    do
+        cxx_entity=${ref};
+    done
+fi
+if test x${cxx_entity} != x; then      # symbolic linked file dereferenced
+    if test ${cxx_entity} = ${CXX}.sh; then
         exit 0
     fi
     if test x${EXEEXT} != x; then
@@ -136,22 +205,24 @@ if test x${sym_link} = xl; then     # symbolic link
     else
         cxx_entity_base=${cxx_entity}
     fi
-    echo "  mv ${cxx_entity} ${cxx_entity_base}_proper${EXEEXT}"
-    mv -f ${cxx_entity} ${cxx_entity_base}_proper${EXEEXT}
 else
-    echo "  mv ${CXX}${EXEEXT} ${CXX}_proper${EXEEXT}"
-    mv -f ${CXX}${EXEEXT} ${CXX}_proper${EXEEXT}
-    cxx_entity=${CXX}${EXEEXT}
+    if test ! ${sys_mingw} || test ! -f ${CXX}_proper${EXEEXT}; then
+        echo "  mv ${CXX}${EXEEXT} ${CXX}_proper${EXEEXT}"
+        mv -f ${CXX}${EXEEXT} ${CXX}_proper${EXEEXT}
+    fi
+    cxx_entity_base=${gcc_path}/${CXX}_proper
 fi
 
 echo '#!/bin/sh' > ${CC}.sh
-echo ${gcc_path}/${c_entity}_proper -no-integrated-cpp '"$@"'   >> ${CC}.sh
+echo ${c_entity_base} -no-integrated-cpp '"$@"'     \
+        >> ${CC}.sh
 echo '#!/bin/sh' > ${CXX}.sh
-echo ${gcc_path}/${cxx_entity}_proper -no-integrated-cpp '"$@"' >> ${CXX}.sh
+echo ${cxx_entity_base} -no-integrated-cpp '"$@"'   \
+        >> ${CXX}.sh
 chmod a+x ${CC}.sh ${CXX}.sh
 
-echo "  ${LN_S} ${CC}.sh ${c_entity}"
-${LN_S} ${CC}.sh ${c_entity}
-echo "  ${LN_S} ${CXX}.sh ${cxx_entity}"
-${LN_S} ${CXX}.sh ${cxx_entity}
+echo "  ${LN_S} ${CC}.sh ${CC}"
+${LN_S} -f ${CC}.sh ${CC}
+echo "  ${LN_S} ${CXX}.sh ${CXX}"
+${LN_S} -f ${CXX}.sh ${CXX}
